Added - Cambio en el _Layout para incluir Header y Footer de Antbox, se agregan estilos y cambio en el index

+++ AntBoxFrontEnd/Services/Zipcodes/ZipCodeResponse.cs
@@ -10,10 +10,10 @@ namespace AntBoxFrontEnd.Services.Zipcodes
     {
 
         [JsonProperty("longitude")]
-        Double Logitude { get; set; }
+        decimal Logitude { get; set; }
 
         [JsonProperty("latitude")]
-        Double Latitude { get; set; }
+        decimal Latitude { get; set; }
 
         [JsonProperty("neighborhood")]
         string Neighborhood { get; set; }
